ForInInit requires a type declaration.
ForInInit introduces a name in the enclosing function's environment.  Without
a type declaration, we can't add it to the environment of the function, since
there is neither a type nor an expression of the type on the RHS.  The RHS is
a collection containing values of the required type.

We'll get into type inference some other time.