All of tourney.tjs is now typed. I got through about 30% of the code in index.html. This typed code WONT WORK, mostly because we don't have varargs and dicts, so I can't type calls to DIV, SPAN, etc. that use both. However, all other code, even though I've had to change it around a bit, should work.
Only two minor haskell code changes:
- fixed op precedence bug in parser. == is higher than &.
- stopped checking for unbound type IDs, since functions inside a forall were not working.

Many comments:
Typing this code required me to look at all uses of a variable, like asgn, to try to piece together what exactly made the object up. Is annoying! That's why we're doing static typing, so other people who look at the code will know what's happening =).

$F returns U(string, Array<string>), but it so often just returns string that nobody checks for the array return type (it only returns an array when you use it on a multi-input select box). Not sure if we can check that statically, but I defined a helper:
function $Fs(n) :: HTMLElement -> string) {
  var res = $F(n);
  if (typeof res == "string")
    return res;
  return res[0];
}

There's a problem with array being invariant. DIV takes U(HTMLElement, string), as a vararg. Convering this to array makes it take Array<U(HTMLElement, string)>. But now, giving it an array of only strings doesn't work, because the type of that array is Array<string>. If this were still varargs, then it would work. I guess there is a difference - the varargs aren't mutable, while the array is.

due to the functional nature of 173tourney, I'm having to write tons of annotations. Example:
    var testSuitesE = responseE.filter_e(isSuccess)
        .lift_e(function(r) { return r.value; });
turns into this monstrosity:
    var testSuitesE = responseE.filter_e@[({success :: bool})](isSuccess)
        .lift_e@[({value :: TestInfoT}), (TestInfoT)](
          function(r) :: ({value :: TestInfoT} -> TestInfoT) { return r.value; });
Maybe we can infer types of anonymous functions, if they are immediately given as arguments to another function? would save a lot of trouble. Also a lot of this would go away with inferring of parapoly.

stuff is not type-safe because EventStreams aren't parametrized. I'm annotating each call to filter_e with what it should be, but the responseE could contain things that are not that.
BUT: This wouldn't type-check if I were actually typing flapjax directly, so it's not a problem with our type system so much. It's more that we're typing "external" things, and the external things would be blamed if they violate the types I gave them.