tourney's index.html is now 50% type-checked. no changes made to haskell type-checker, yay. it's annoying to do this, though, mostly because each anonymous function must be typed (doesn't TS infer some lambda types?) and because I'm lazy and haven't done parapoly inference yet. more notes follow.
formatCurrentTest seemed to use an array as a tuple. i changed it to use an object insteead. but, it might be interesting to support them.

the app would be much better specified with types everywhere. now the only way to know what is being sent where is to look at the haskell server code, and even there, the types of the objects aren't well specified, i don't think.

i've been saying we don't have dicts, so we can't do some stuff. but actually, what's being passed into DIV et. al. isn't a dict, but an object full of optional fields. If a field isn't specified, it's set to undefined. Maybe we should change object subtyping to take this into account. or, (probably better), we can detect when this happens and automatically fill in objects with fields set to undefined.

YES! Found a bug!! asgnSolution changed "result.value" to an error message, when result is just a DIV and has no value field. It should be report(msg, "Retrieving solution") or gold.value or result.innerHTML or something.

we should really make parens required around funcs. it'll make other syntax easier. also it'll remove ambiguity when functions don't return anything. That is, this:
function (z) :: int -> {
  return 4;
}
gives an error since it thinks you're trying to define the object type that you're returning.

This code doesn't work:
  function () :: (-> (-> HTMLElement)) {
    var resultDiv :: HTMLElement?
    if (pred) { resultDiv = HE; }
    else      { resultDiv = HE; }
    return function() :: (-> HTMLElement) { return resultDiv; };
  };
because resultDiv is still U(..., undefined) in the func body. should be some way to occurrence type that away... or, the code could just be re-structured, i 'sppose.

calling EventStream.lift_e changes the parametrized type of EventStream... this might lead to problems.