10% of flapjax is typed. read on for comments.
we have a problem w/ closures, kind of. for example, see like 413 of
flapjax.tjs . updater is a function?. we know it will never change, so
the refinement in the Behavior constructor should really carry over
into the closure. in this case, we can refactor the code (although
it's a bit slower), so that's fine.

but not always. look at bindE, line 466. the inner function should
really be refining prevE, and then assigning to it. however, we can't
assign to a 'global' union, so the code doesn't work. we need some way
of telling when a 'global' union will never be modified again, meaning
that it's ok to mutate to it and assign to it.

one other thing. look at receiverE, line 439. it adds a 'sendEvent'
field to an eventstream. i take this to mean that some, but not all,
eventstreams have a sendEvent, so I added sendEvent as a function? to
the type. however, that means that we can never do something like
"evt.sendEvent = function () ..." because that'd be mutating a
union. =(.