Interesting development!
The following code was failing to type check:
external obj1 :: {foo :: U(undefined, int)};
external obj2 :: {foo :: undefined};
obj1 = obj2;

The reason?
while supertyping field foo:
type int in the union 'U (undefined, int)' isn't a subtype of the rhs, undefined:
case fall-through. TId "int" is not a subtype of TId "undefined")

The type-checker thought that the union field was writeable / mutable, so it was
trying to super-type the field. ive changed it so that union fields are automatically
not writeable, but this is an assumption that is guaranteed elsewhere.