Dataflow analysis uses more static type information.
There are various situations where the dataflow analysis concludes that the
type of a variable at a statement (in ANF) is indeterminate.  Previously, the
analysis would set the type to TUnk.  However, if we know that the variable
has a declared type T, then we set it to runtime(T), which is usually better
than TUnk.